refactor: Move Text and Title components to correct directories
- Relocated Text and Title components to their appropriate project directories
- Updated component imports and file structure
