+++ src/components/title/index.ts
@@ -1,1 +1,1 @@
-export { default as Text } from './Text.tsx'
+export { default as Title } from './Title.tsx'
